regmap: rbtree: Simplify adjacent node look-up
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Lars-Peter Clausen <lars@metafoo.de>
commit 194c753a214ba7f1497552dd530021884d164146

A register which is adjacent to a node will either be left to the first
register or right to the last register. It will not be within the node's range,
so there is no point in checking for each register cached by the node whether
the new register is next to it. It is sufficient to check whether the register
comes before the first register or after the last register of the node.

	Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
	Signed-off-by: Mark Brown <broonie@linaro.org>
(cherry picked from commit 194c753a214ba7f1497552dd530021884d164146)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>