i2c: add quirk structure to describe adapter flaws
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
Rebuild_CHGLOG: - [i2c] add quirk structure to describe adapter flaws (Gustavo Duarte) [1221075]
Rebuild_FUZZ: 94.74%
commit-author Wolfram Sang <wsa@the-dreams.de>
commit 2187f03a9576c4fb8342deed912b5ba6fcf98cba

The number of I2C adapters which are not fully I2C compatible is rising,
sadly. Drivers usually do handle the flaws, still the user receives only
some errno for a transfer which normally can be expected to work. This
patch introduces a formal description of flaws. One advantage is that
the core can check before the actual transfer if the messages could be
transferred at all. This is done in the next patch. Another advantage is
that we can pass this information to the user so the restrictions are
exactly known and further actions can be based on that. This will be
done later after some stabilization period for this description.

	Signed-off-by: Wolfram Sang <wsa@the-dreams.de>
	Tested-by: Ray Jui <rjui@broadcom.com>
	Tested-by: Ivan T. Ivanov <iivanov@mm-sol.com>
	Tested-by: Neelesh Gupta <neelegup@linux.vnet.ibm.com>
Tested-By: Ludovic Desroches <ludovic.desroches@atmel.com>
(cherry picked from commit 2187f03a9576c4fb8342deed912b5ba6fcf98cba)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>