xen-blkfront: remove type check from blkfront_setup_discard
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Olaf Hering <olaf@aepfle.de>
commit 1c8cad6c1bbcacc07332539f2d7dee22521f9ae0

In its initial implementation a check for "type" was added, but only phy
and file are handled. This breaks advertised discard support for other
type values such as qdisk.

Fix and simplify this function: If the backend advertises discard
support it is supposed to implement it properly, so enable
feature_discard unconditionally. If the backend advertises the need for
a certain granularity and alignment then propagate both properties to
the blocklayer. The discard-secure property is a boolean, update the code
to reflect that.

	Signed-off-by: Olaf Hering <olaf@aepfle.de>
	Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
(cherry picked from commit 1c8cad6c1bbcacc07332539f2d7dee22521f9ae0)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>