fm10k: Add support for ptp to hw specific files
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Alexander Duyck <alexander.h.duyck@intel.com>
commit 5f226ddb5b0c477bd512085b0b1d1052a24f0020

This change adds the messaging support needed to support PTP.  In the case
of Tx timestamps it is necessary for the Switch Management entity to return
the frames via the mailbox as the host interface cannot know which port the
timestamp will be delivered to.  In addition there is only one clock on the
entire switch, as such the entity that has BAR 4 access is the only one who
can actually update the frequency as it is the only one with access.

	Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
	Acked-by: Richard Cochran <richardcochran@gmail.com>
	Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 5f226ddb5b0c477bd512085b0b1d1052a24f0020)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>