Drivers: hv: hv_balloon: do not online pages in offline blocks
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
Rebuild_CHGLOG: - [hv] hv_balloon: do not online pages in offline blocks (Vitaly Kuznetsov) [1133403 1218868]
Rebuild_FUZZ: 88.29%
commit-author Vitaly Kuznetsov <vkuznets@redhat.com>
commit d6cbd2c3a3db437708520c66f285c69ef028ac1f

Currently we add memory in 128Mb blocks but the request from host can be
aligned differently. In such case we add a partially backed block and
when this block goes online we skip onlining pages which are not backed
(hv_online_page() callback serves this purpose). When we receive next
request for the same host add region we online pages which were not backed
before with hv_bring_pgs_online(). However, we don't check if the the block
in question was onlined and online this tail unconditionally. This is bad as
we avoid all online_pages() logic: these pages are not accounted, we don't
send notifications (and hv_balloon is not the only receiver of them),...
And, first of all, nobody asked as to online these pages. Solve the issue by
checking if the last previously backed page was onlined and onlining the tail
only in case it was.

	Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
	Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit d6cbd2c3a3db437708520c66f285c69ef028ac1f)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>