openvswitch: Create right mask with disabled megaflows
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Pravin B Shelar <pshelar@nicira.com>
commit f47de068f68db91b89e0d3335230d07e02da8727

If megaflows are disabled, the userspace does not send the netlink attribute
OVS_FLOW_ATTR_MASK, and the kernel must create an exact match mask.

sw_flow_mask_set() sets every bytes (in 'range') of the mask to 0xff, even the
bytes that represent padding for struct sw_flow, or the bytes that represent
fields that may not be set during ovs_flow_extract().
This is a problem, because when we extract a flow from a packet,
we do not memset() anymore the struct sw_flow to 0.

This commit gets rid of sw_flow_mask_set() and introduces mask_set_nlattr(),
which operates on the netlink attributes rather than on the mask key. Using
this approach we are sure that only the bytes that the user provided in the
flow are matched.

Also, if the parse_flow_mask_nlattrs() for the mask ENCAP attribute fails, we
now return with an error.

This bug is introduced by commit 0714812134d7dcadeb7ecfbfeb18788aa7e1eaac
("openvswitch: Eliminate memset() from flow_extract").

	Reported-by: Alex Wang <alexw@nicira.com>
	Signed-off-by: Daniele Di Proietto <ddiproietto@vmware.com>
	Signed-off-by: Andy Zhou <azhou@nicira.com>
	Signed-off-by: Pravin B Shelar <pshelar@nicira.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit f47de068f68db91b89e0d3335230d07e02da8727)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>