USB: xhci: Reset a halted endpoint immediately when we encounter a stall.
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
Rebuild_CHGLOG: - [usb] xhci: Reset a halted endpoint immediately when we encounter a stall (Don Zickus) [1232920]
Rebuild_FUZZ: 95.71%
commit-author Mathias Nyman <mathias.nyman@linux.intel.com>
commit 8e71a322fdb127814bcba423a512914ca5bc6cf5

If a device is halted and reuturns a STALL, then the halted endpoint
needs to be cleared both on the host and device side. The host
side halt is cleared by issueing a xhci reset endpoint command. The device side
is cleared with a ClearFeature(ENDPOINT_HALT) request, which should
be issued by the device driver if a URB reruen -EPIPE.

Previously we cleared the host side halt after the device side was cleared.
To make sure the host side halt is cleared in time we want to issue the
reset endpoint command immedialtely when a STALL status is encountered.

Otherwise we end up not following the specs and not returning -EPIPE
several times in a row when trying to transfer data to a halted endpoint.

Fixes: bcef3fd (USB: xhci: Handle errors that cause endpoint halts.)
	Cc: <stable@vger.kernel.org> # v2.6.33+
	Tested-by: Felipe Balbi <balbi@ti.com>
	Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 8e71a322fdb127814bcba423a512914ca5bc6cf5)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>