cxl: Add base builtin support
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Ian Munsie <imunsie@au1.ibm.com>
commit 10542ca0156f60571ef41799d44d40dd4cb0a473
Empty-Commit: Cherry-Pick Conflicts during history rebuild.
Will be included in final tarball splat. Ref for failed cherry-pick at:
ciq/ciq_backports/kernel-3.10.0-327.el7/10542ca0.failed

This adds the base cxl support that cannot be built as a module. Specifically
it adds the cxl callbacks that are called from the core powerpc mm code which
must always exist irrespective of if the cxl module is loaded or not. This is
similar to how cell works with CONFIG_SPU_BASE.

This adds a cxl_slbia() call (similar to spu_flush_all_slbs()) which checks if
the cxl module is loaded and in use, returning immediately if it is not. If it
is in use it calls into the cxl SLB invalidation code.

	Signed-off-by: Ian Munsie <imunsie@au1.ibm.com>
	Signed-off-by: Michael Neuling <mikey@neuling.org>
	Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
(cherry picked from commit 10542ca0156f60571ef41799d44d40dd4cb0a473)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>

# Conflicts:
#	drivers/misc/Kconfig
#	drivers/misc/Makefile