cxl: Change contexts_lock to a mutex to fix sleep while atomic bug
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Ian Munsie <imunsie@au1.ibm.com>
commit ee41d11d53c8fc4968f0816504651541d606cf40

We had a known sleep while atomic bug if a CXL device was forcefully
unbound while it was in use. This could occur as a result of EEH, or
manually induced with something like this while the device was in use:

echo 0000:01:00.0 > /sys/bus/pci/drivers/cxl-pci/unbind

The issue was that in this code path we iterated over each context and
forcefully detached it with the contexts_lock spin lock held, however
the detach also needed to take the spu_mutex, and call schedule.

This patch changes the contexts_lock to a mutex so that we are not in
atomic context while doing the detach, thereby avoiding the sleep while
atomic.

Also delete the related TODO comment, which suggested an alternate
solution which turned out to not be workable.

	Cc: stable@vger.kernel.org
	Signed-off-by: Ian Munsie <imunsie@au1.ibm.com>
	Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
(cherry picked from commit ee41d11d53c8fc4968f0816504651541d606cf40)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>