NFS: fix subtle change in COMMIT behavior
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Weston Andros Adamson <dros@primarydata.com>
commit cb1410c71e0b6b2eba8c1890645a76ff86169d24

Recent work in the pgio layer made it possible for there to be more than one
request per page. This caused a subtle change in commit behavior, because
write.c:nfs_commit_unstable_pages compares the number of *pages* waiting for
writeback against the number of requests on a commit list to choose when to
send a COMMIT in a non-blocking flush.

This is probably hard to hit in normal operation - you have to be using
rsize/wsize < PAGE_SIZE, or pnfs with lots of boundaries that are not page
aligned to have a noticeable change in behavior.

	Signed-off-by: Weston Andros Adamson <dros@primarydata.com>
	Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
(cherry picked from commit cb1410c71e0b6b2eba8c1890645a76ff86169d24)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>