bnx2x: Fix linearization for encapsulated packets
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Yuval Mintz <Yuval.Mintz@qlogic.com>
commit 592b9b8d687827086344705631494e5f232bee50

Due to FW constraints, driver must make sure that transmitted SKBs will
not be too fragmented, or in the case that they are - that each 'window'
of fragments passed to the FW would contain at least an mss worth of data.

For encapsultaed packets the calculation is wrong, since it ignores the
inner headers in the calculation of the headers' length.
This could lead to a FW assertion in case of a too-fragmented encapsulated
packet.

	Signed-off-by: Yuval Mintz <Yuval.Mintz@qlogic.com>
	Signed-off-by: Ariel Elior <Ariel.Elior@qlogic.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 592b9b8d687827086344705631494e5f232bee50)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>