hrtimer: Store cpu-number in struct hrtimer_cpu_base
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Viresh Kumar <viresh.kumar@linaro.org>
commit cddd02489f52ccf635ed65931214729a23b93cd6

In lowres mode, hrtimers are serviced by the tick instead of a clock
event. Now it works well as long as the tick stays periodic but we
must also make sure that the hrtimers are serviced in dynticks mode.

Part of that job consist in kicking a dynticks hrtimer target in order
to make it reconsider the next tick to schedule to correctly handle the
hrtimer's expiring time. And that part isn't handled by the hrtimers
subsystem.

To prepare for fixing this, we need __hrtimer_start_range_ns() to be
able to resolve the CPU target associated to a hrtimer's object
'cpu_base' so that the kick can be centralized there.

So lets store it in the 'struct hrtimer_cpu_base' to resolve the CPU
without overhead. It is set once at CPU's online notification.

	Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
	Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>
Link: http://lkml.kernel.org/r/1403393357-2070-4-git-send-email-fweisbec@gmail.com
	Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
(cherry picked from commit cddd02489f52ccf635ed65931214729a23b93cd6)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>