iSCSI: let session recovery_tmo sysfs writes persist across recovery
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Chris Leech <cleech@redhat.com>
commit 9c8108a4d3a837c51a29f28229a06d97654eaeb6

The iSCSI session recovery_tmo setting is writeable in sysfs, but it's
also set every time a connection is established when parameters are set
from iscsid over netlink.  That results in the timeout being reset to
the default value after every recovery.

The DM multipath tools want to use the sysfs interface to lower the
default timeout when there are multiple paths to fail over.  It has
caused confusion that we have a writeable sysfs value that seem to keep
resetting itself.

This patch adds an in-kernel flag that gets set once a sysfs write
occurs, and then ignores netlink parameter setting once it's been
modified via the sysfs interface.  My thinking here is that the sysfs
interface is much simpler for external tools to influence the session
timeout, but if we're going to allow it to be modified directly we
should ensure that setting is maintained.

	Signed-off-by: Chris Leech <cleech@redhat.com>
	Reviewed-by: Mike Christie <michaelc@cs.wisc.edu>
	Signed-off-by: James Bottomley <JBottomley@Odin.com>
(cherry picked from commit 9c8108a4d3a837c51a29f28229a06d97654eaeb6)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>