sfc: Report TX completions to BQL after all TX events in interrupt
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-327.el7
commit-author Peter Dunning <pdunning@solarflare.com>
commit c936835c1ec6f871f32c9b87a7708700320075b3

The limit for BQL is updated each time we call
netdev_tx_completed_queue.
Without this patch the BQL limit was updated for every TX event we
see.
The issue was that this only updated the limit to handle the data
we complete in two events as the first event wouldn't show that
enough traffic had been processed between them.

This was OK when interrupt moderation was off but not when it was
on as more data had to be completed in a single interrupt.

The patch changes this so that we do report the completion to BQL
only when all the TX events in the interrupt have been processed.

	Signed-off-by: Shradha Shah <sshah@solarflare.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit c936835c1ec6f871f32c9b87a7708700320075b3)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>